requirements.txt
Pillow==8.4.0
numpy==1.26.4
pandas==2.2.1
spotipy==2.23.0
streamlit-autorefresh==1.0.1
gspread==5.7.2
oauth2client==1.2.0
pylast==5.3.0
tzlocal==5.2
plotly==5.24.1
google-generativeai